fix: type error
