+++ src/services/recruit/apis/index.ts
@@ -5,6 +5,7 @@ export * from './completeRecruit';
 export * from './constants';
 export * from './createRecruit';
 export * from './getMyRecruitApplication';
+export * from './getRecruitApplicants';
 export * from './getRecruitApplication';
 export * from './getRecruitDetail';
 export * from './getRecruitParticipants';
