+++ src/services/recruit/hooks/useRecruitApplicants.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 
 import { queryKeys } from '~/react-query/common';
-import { getRecruitApplicants } from '~/services/recruit/apis2';
+import { getRecruitApplicants } from '~/services/recruit/apis';
 import { toMs } from '~/utils';
 
 export const useRecruitApplicants = (recruitId: number) => {
