+++ src/services/recruit/index.ts
@@ -1,4 +1,3 @@
-export * from './apis2';
 export * from './utils';
 export * from './hooks';
 export * from './apis';
