+++ src/services/recruit/utils/index.ts
@@ -1,5 +1,4 @@
 export * from './types';
 export * from './getRecruitThemeByCategory';
 export * from './getDisplayCategoryName';
-export * from './recruitTypeConvertor';
 export * from './reconfirmRecruitFormUnload';
